Add stopexec
stopexec allows one to dump a checkpoint of a process just before it
starts, capturing command line arguments, environment variables, and
inherited file descriptors. While this can be done with a sufficiently
sophisticated shell script, putting this data in CRIU format allows
the start of a program to be treated in the same manner as resuming at
all other points in an applications execution must be treated--by
dumping and restoring a checkpoint.

Signed-off-by: Christopher Covington <cov@codeaurora.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>