Add ptrace-wait
ptrace-wait allows the user to wait for any PID, regardless of
parent-child relationship, to exit, by using ptrace. Furthermore,
using perf_event_open, it allows the user to wait a specified number
of instructions rather than the full length. This can be used, for
example, to dump a checkpoint every N instructions in a program
execution, and to run each restored checkpoint for N instructions.
Given a fast system on which the dumps are performed and one or more
slow system(s) on which the restores are performed, this enables
"fast forwarding" through an application's execution.

Signed-off-by: Christopher Covington <cov@codeaurora.org>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>